chore: Move fingerprints to the correct directory and use object fingerprints
