+++ patches/src/main/kotlin/app/morphe/patches/youtube/interaction/seekbar/Fingerprints.kt
@@ -22,6 +22,7 @@ import app.morphe.patcher.opcode
 import app.morphe.patcher.string
 import app.morphe.patches.all.misc.resources.ResourceType
 import app.morphe.patches.all.misc.resources.resourceLiteral
+import app.morphe.patches.shared.FormatStreamModelToStringFingerprint
 import app.morphe.patches.youtube.shared.SeekbarFingerprint
 import app.morphe.patches.youtube.shared.VideoStreamingDataToStringFingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
@@ -156,17 +157,10 @@ internal object VideoStreamingDataAllowSeekingFingerprint : Fingerprint(
     )
 )
 
-private object FormatStreamModelClassFingerprint : Fingerprint(
-    returnType = "Ljava/lang/String;",
-    filters = listOf(
-        string("FormatStream(itag=")
-    )
-)
-
 // DVR window duration in seconds; 0 for non-DVR streams.
 // Caller multiplies result by 1e6 with 4-hour fallback when <= 0, logs "windowMaxMediaTimeUs".
 internal object FormatStreamModelMaxDVRDurationFingerprint : Fingerprint(
-    classFingerprint = FormatStreamModelClassFingerprint,
+    classFingerprint = FormatStreamModelToStringFingerprint,
     accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.FINAL),
     returnType = "D",
     parameters = listOf(),
