+++ patches/src/main/kotlin/app/morphe/patches/youtube/video/information/Fingerprints.kt
@@ -27,7 +27,6 @@ import app.morphe.patches.youtube.shared.PlaybackSpeedOnItemClickParentFingerpri
 import app.morphe.patches.youtube.shared.VideoQualityChangedFingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
 import com.android.tools.smali.dexlib2.Opcode
-import com.sun.tools.javac.main.Option
 
 internal object PlaybackSpeedOnItemClickFingerprint : Fingerprint(
     classFingerprint = PlaybackSpeedOnItemClickParentFingerprint,
@@ -319,6 +318,17 @@ internal fun getPlaybackParametersSetterFingerprint(playbackParametersType: Stri
         methodDef.implementation != null
             && classDef.interfaces.contains("Landroidx/media3/exoplayer/ExoPlayer;")
     }
+) {}
+
+internal fun getExoPlayerImplFingerprint(playbackParametersType: String) = object : Fingerprint(
+    classFingerprint = getPlaybackParametersSetterFingerprint(playbackParametersType),
+    name = "<init>",
+    filters = listOf(
+        methodCall(
+            opcode = Opcode.INVOKE_DIRECT,
+            name = "<init>"
+        )
+    )
 ) {}
 
 /**
